modified structure
