+++ web/package.json
@@ -18,6 +18,7 @@
     "petra-plugin-wallet-adapter": "^0.3.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-icons": "^4.11.0",
     "react-router-dom": "^6.17.0",
     "react-scripts": "5.0.1",
     "typescript": "^4.9.5",
